+++ packages/annotation-converter/src/converter.ts
@@ -37,12 +37,10 @@ import type {
     Singleton,
     TypeDefinition
 } from '@sap-ux/vocabularies-types';
-import type { ReferencesWithMap } from './utils';
 import {
     addGetByValue,
     alias,
     Decimal,
-    defaultReferences,
     EnumIsFlag,
     lazy,
     splitAtFirst,
@@ -52,6 +50,7 @@ import {
     TermToTypes,
     unalias
 } from './utils';
+import { VocabularyReferences } from '@sap-ux/vocabularies-types/vocabularies/VocabularyReferences';
 
 /**
  * Symbol to extend an annotation with the reference to its target.
@@ -438,12 +437,13 @@ function parseValue(
         case 'Date':
             return propertyValue.Date;
         case 'EnumMember':
-            const aliasedEnum = converter.alias(propertyValue.EnumMember);
-            const splitEnum = aliasedEnum.split(' ');
-            if (splitEnum[0] && EnumIsFlag[substringBeforeFirst(splitEnum[0], '/')]) {
+            const splitEnum = propertyValue.EnumMember.split(' ').map((enumValue) =>
+                converter.toDefaultAlias(enumValue)
+            );
+            if (splitEnum[0] !== undefined && EnumIsFlag[substringBeforeFirst(splitEnum[0], '/')]) {
                 return splitEnum;
             }
-            return aliasedEnum;
+            return splitEnum[0];
 
         case 'PropertyPath':
             return {
@@ -851,17 +851,6 @@ function isV4NavigationProperty(
     return !!(navProp as BaseNavigationProperty).targetTypeName;
 }
 
-/**
- * Split the alias from the term value.
- *
- * @param references the current set of references
- * @param termValue the value of the term
- * @returns the term alias and the actual term value
- */
-function splitTerm(references: ReferencesWithMap, termValue: string) {
-    return splitAtLast(alias(references, termValue), '.');
-}
-
 function convertAnnotation(converter: Converter, target: any, rawAnnotation: RawAnnotation): Annotation {
     let annotation: any;
     if (rawAnnotation.record) {
@@ -1133,12 +1122,20 @@ class Converter {
         this.convertedOutput.diagnostics.push({ message });
     }
 
+    /**
+     * Split the alias from the term value.
+     *
+     * @param term the value of the term
+     * @returns the term alias and the actual term value
+     */
     splitTerm(term: string) {
-        return splitTerm(this.rawMetadata.references, term);
+        const aliased = alias(VocabularyReferences, term);
+        return splitAtLast(aliased, '.');
     }
 
-    alias(value: string) {
-        return alias(this.rawMetadata.references, value);
+    toDefaultAlias(value: string) {
+        const unaliased = unalias(this.rawMetadata.references, value) ?? '';
+        return alias(VocabularyReferences, unaliased);
     }
 
     unalias(value: string | undefined) {
@@ -1556,19 +1553,19 @@ function convertTypeDefinition(converter: Converter, rawTypeDefinition: RawTypeD
  * @returns the converted representation of the metadata.
  */
 export function convert(rawMetadata: RawMetadata): ConvertedMetadata {
-    // fall back on the default references if the caller does not specify any
-    if (rawMetadata.references.length === 0) {
-        rawMetadata.references = defaultReferences;
-    }
-
     // Converter Output
     const convertedOutput: ConvertedMetadata = {
         version: rawMetadata.version,
         namespace: rawMetadata.schema.namespace,
         annotations: rawMetadata.schema.annotations,
-        references: defaultReferences.concat(rawMetadata.references),
+        references: VocabularyReferences.concat(rawMetadata.references),
         diagnostics: []
     } as any;
+
+    // fall back on the default references if the caller does not specify any
+    if (rawMetadata.references.length === 0) {
+        rawMetadata.references = VocabularyReferences;
+    }
 
     // Converter
     const converter = new Converter(rawMetadata, convertedOutput);