+++ packages/annotation-converter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/annotation-converter
 
+## 0.6.6
+
+### Patch Changes
+
+-   9be80ce: Consider the alias when converting annotations
+
+## 0.6.5
+
+### Patch Changes
+
+-   217e517: The vocabulary types now provide a complete set of vocabulary references in constant `VocabularyReferences`
+
 ## 0.6.4
 
 ### Patch Changes