+++ packages/fe-mockserver-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/fe-mockserver-core
 
+## 1.1.67
+
+### Patch Changes
+
+-   Updated dependencies [9be80ce]
+    -   @sap-ux/annotation-converter@0.6.6
+
+## 1.1.66
+
+### Patch Changes
+
+-   Updated dependencies [217e517]
+    -   @sap-ux/annotation-converter@0.6.5
+    -   @sap-ux/edmx-parser@0.5.13
+
 ## 1.1.65
 
 ### Patch Changes